+++ multisig/mandos/reject_elrond_to_ethereum_tx_batch.scen.json
@@ -54,7 +54,7 @@
                     "balance": "*",
                     "storage": {
                         "str:action_data.item|u32:1": {
-                            "1-action_type": "u8:6",
+                            "1-action_type": "u8:1",
                             "2-batch_id": "u64:1",
                             "3-tx_batch_status_len": "u32:2",
                             "4-tx_batch_status_vec": "u8:4|u8:4"
