+++ multisig/src/setup.rs
@@ -1,9 +1,9 @@
 elrond_wasm::imports!();
 
-use crate::user_role::UserRole;
 use eth_address::EthAddress;
 
 use fee_estimator_module::ProxyTrait as _;
+use multi_transfer_esdt::ProxyTrait as _;
 use token_module::ProxyTrait as _;
 use tx_batch_module::ProxyTrait as _;
 
@@ -53,14 +53,14 @@ pub trait SetupModule:
 
     #[only_owner]
     #[endpoint(addBoardMember)]
-    fn add_board_member(&self, board_member: ManagedAddress) {
-        self.change_user_role(board_member, UserRole::BoardMember);
+    fn add_board_member_endpoint(&self, board_member: ManagedAddress) {
+        self.add_board_member(&board_member);
     }
 
     #[only_owner]
     #[endpoint(removeUser)]
-    fn remove_user(&self, user: ManagedAddress) {
-        self.change_user_role(user, UserRole::None);
+    fn remove_user(&self, board_member: ManagedAddress) {
+        self.remove_board_member(&board_member);
         let num_board_members = self.num_board_members().get();
         require!(num_board_members > 0, "cannot remove all board members");
         require!(
@@ -140,7 +140,7 @@ pub trait SetupModule:
     #[only_owner]
     #[endpoint(changeFeeEstimatorContractAddress)]
     fn change_fee_estimator_contract_address(&self, new_address: ManagedAddress) {
-        self.setup_get_esdt_safe_proxy_instance()
+        self.get_esdt_safe_proxy_instance()
             .set_fee_estimator_contract_address(new_address)
             .execute_on_dest_context();
     }
@@ -148,7 +148,7 @@ pub trait SetupModule:
     #[only_owner]
     #[endpoint(changeElrondToEthGasLimit)]
     fn change_elrond_to_eth_gas_limit(&self, new_gas_limit: BigUint) {
-        self.setup_get_esdt_safe_proxy_instance()
+        self.get_esdt_safe_proxy_instance()
             .set_eth_tx_gas_limit(new_gas_limit)
             .execute_on_dest_context();
     }
@@ -156,7 +156,7 @@ pub trait SetupModule:
     #[only_owner]
     #[endpoint(changeDefaultPricePerGasUnit)]
     fn change_default_price_per_gas_unit(&self, token_id: TokenIdentifier, new_value: BigUint) {
-        self.setup_get_esdt_safe_proxy_instance()
+        self.get_esdt_safe_proxy_instance()
             .set_default_price_per_gas_unit(token_id, new_value)
             .execute_on_dest_context();
     }
@@ -164,7 +164,7 @@ pub trait SetupModule:
     #[only_owner]
     #[endpoint(changeTokenTicker)]
     fn change_token_ticker(&self, token_id: TokenIdentifier, new_ticker: ManagedBuffer) {
-        self.setup_get_esdt_safe_proxy_instance()
+        self.get_esdt_safe_proxy_instance()
             .set_token_ticker(token_id, new_ticker)
             .execute_on_dest_context();
     }
@@ -177,7 +177,7 @@ pub trait SetupModule:
         ticker: ManagedBuffer,
         #[var_args] opt_default_value_in_dollars: OptionalValue<BigUint>,
     ) {
-        self.setup_get_esdt_safe_proxy_instance()
+        self.get_esdt_safe_proxy_instance()
             .add_token_to_whitelist(token_id, ticker, opt_default_value_in_dollars)
             .execute_on_dest_context();
     }
@@ -185,7 +185,7 @@ pub trait SetupModule:
     #[only_owner]
     #[endpoint(esdtSafeRemoveTokenFromWhitelist)]
     fn esdt_safe_remove_token_from_whitelist(&self, token_id: TokenIdentifier) {
-        self.setup_get_esdt_safe_proxy_instance()
+        self.get_esdt_safe_proxy_instance()
             .remove_token_from_whitelist(token_id)
             .execute_on_dest_context();
     }
@@ -193,7 +193,7 @@ pub trait SetupModule:
     #[only_owner]
     #[endpoint(esdtSafeSetMaxTxBatchSize)]
     fn esdt_safe_set_max_tx_batch_size(&self, new_max_tx_batch_size: usize) {
-        self.setup_get_esdt_safe_proxy_instance()
+        self.get_esdt_safe_proxy_instance()
             .set_max_tx_batch_size(new_max_tx_batch_size)
             .execute_on_dest_context();
     }
@@ -201,7 +201,7 @@ pub trait SetupModule:
     #[only_owner]
     #[endpoint(esdtSafeSetMaxTxBatchBlockDuration)]
     fn esdt_safe_set_max_tx_batch_block_duration(&self, new_max_tx_batch_block_duration: u64) {
-        self.setup_get_esdt_safe_proxy_instance()
+        self.get_esdt_safe_proxy_instance()
             .set_max_tx_batch_block_duration(new_max_tx_batch_block_duration)
             .execute_on_dest_context();
     }
@@ -209,7 +209,7 @@ pub trait SetupModule:
     #[only_owner]
     #[endpoint(multiTransferEsdtSetMaxRefundTxBatchSize)]
     fn multi_transfer_esdt_set_max_refund_tx_batch_size(&self, new_max_tx_batch_size: usize) {
-        self.setup_get_multi_transfer_esdt_proxy_instance()
+        self.get_multi_transfer_esdt_proxy_instance()
             .set_max_tx_batch_size(new_max_tx_batch_size)
             .execute_on_dest_context();
     }
@@ -220,7 +220,7 @@ pub trait SetupModule:
         &self,
         new_max_tx_batch_block_duration: u64,
     ) {
-        self.setup_get_multi_transfer_esdt_proxy_instance()
+        self.get_multi_transfer_esdt_proxy_instance()
             .set_max_tx_batch_block_duration(new_max_tx_batch_block_duration)
             .execute_on_dest_context();
     }
@@ -231,29 +231,8 @@ pub trait SetupModule:
         &self,
         #[var_args] opt_wrapping_contract_address: OptionalValue<ManagedAddress>,
     ) {
-        self.setup_get_multi_transfer_esdt_proxy_instance()
+        self.get_multi_transfer_esdt_proxy_instance()
             .set_wrapping_contract_address(opt_wrapping_contract_address)
             .execute_on_dest_context();
-    }
-
-    // proxies
-
-    #[proxy]
-    fn setup_esdt_safe_proxy(&self, sc_address: ManagedAddress) -> esdt_safe::Proxy<Self::Api>;
-
-    #[proxy]
-    fn setup_multi_transfer_esdt_proxy(
-        &self,
-        sc_address: ManagedAddress,
-    ) -> multi_transfer_esdt::Proxy<Self::Api>;
-
-    fn setup_get_esdt_safe_proxy_instance(&self) -> esdt_safe::Proxy<Self::Api> {
-        self.setup_esdt_safe_proxy(self.esdt_safe_address().get())
-    }
-
-    fn setup_get_multi_transfer_esdt_proxy_instance(
-        &self,
-    ) -> multi_transfer_esdt::Proxy<Self::Api> {
-        self.setup_multi_transfer_esdt_proxy(self.multi_transfer_esdt_address().get())
     }
 }
