+++ multisig/src/user_role.rs
@@ -3,7 +3,6 @@ elrond_wasm::derive_imports!();
 #[derive(TopEncode, TopDecode, TypeAbi, Clone, Copy, PartialEq)]
 pub enum UserRole {
     None,
-    _Proposer,
     BoardMember,
 }
