+++ multisig/src/util.rs
@@ -3,7 +3,6 @@ use elrond_wasm::elrond_codec::TopEncode;
 
 use transaction::{EthTransaction, EthTxAsMultiValue};
 
-use crate::action::Action;
 use crate::storage::EthBatchHash;
 use crate::user_role::UserRole;
 
@@ -18,24 +17,7 @@ const MAX_BUFFER_SIZE: usize = AVERAGE_SERIALIZED_TX_SIZE * 100;
 
 #[elrond_wasm::module]
 pub trait UtilModule: crate::storage::StorageModule {
-    /// Returns `true` (`1`) if the user has signed the action.
-    /// Does not check whether or not the user is still a board member and the signature valid.
-    #[view]
-    fn signed(&self, user: ManagedAddress, action_id: usize) -> bool {
-        let user_id = self.user_mapper().get_user_id(&user);
-        if user_id == 0 {
-            false
-        } else {
-            self.action_signer_ids(action_id).contains(&user_id)
-        }
-    }
-
-    /// Indicates user rights.
-    /// `0` = no rights,
-    /// `1` = can propose, but not sign,
-    /// `2` = can propose and sign.
-    #[view(userRole)]
-    fn user_role(&self, user: &ManagedAddress) -> UserRole {
+    fn get_user_role(&self, user: &ManagedAddress) -> UserRole {
         let user_id = self.user_mapper().get_user_id(user);
         if user_id == 0 {
             UserRole::None
@@ -44,68 +26,11 @@ pub trait UtilModule: crate::storage::StorageModule {
         }
     }
 
-    /// Lists all users that can sign actions.
-    #[view(getAllBoardMembers)]
-    fn get_all_board_members(&self) -> MultiValueEncoded<ManagedAddress> {
-        self.get_all_users_with_role(UserRole::BoardMember)
-    }
+    fn is_valid_action_id(&self, action_id: usize) -> bool {
+        let min_id = 1;
+        let max_id = self.action_mapper().len();
 
-    #[view(getAllStakedRelayers)]
-    fn get_all_staked_relayers(&self) -> MultiValueEncoded<ManagedAddress> {
-        let relayers = self.get_all_board_members().to_vec();
-        let mut staked_relayers = ManagedVec::new();
-
-        for relayer in &relayers {
-            if self.has_enough_stake(&relayer) {
-                staked_relayers.push(relayer);
-            }
-        }
-
-        staked_relayers.into()
-    }
-
-    /// Gets addresses of all users who signed an action and are still board members.
-    /// All these signatures are currently valid.
-    #[view(getActionSignerCount)]
-    fn get_action_signer_count(&self, action_id: usize) -> usize {
-        self.action_signer_ids(action_id).len()
-    }
-
-    /// It is possible for board members to lose their role.
-    /// They are not automatically removed from all actions when doing so,
-    /// therefore the contract needs to re-check every time when actions are performed.
-    /// This function is used to validate the signers before performing an action.
-    /// It also makes it easy to check before performing an action.
-    #[view(getActionValidSignerCount)]
-    fn get_action_valid_signer_count(&self, action_id: usize) -> usize {
-        self.action_signer_ids(action_id)
-            .iter()
-            .filter(|signer_id| {
-                let signer_role = self.user_id_to_role(*signer_id).get();
-                signer_role.is_board_member()
-            })
-            .count()
-    }
-
-    /// Returns `true` (`1`) if `getActionValidSignerCount >= getQuorum`.
-    #[view(quorumReached)]
-    fn quorum_reached(&self, action_id: usize) -> bool {
-        let quorum = self.quorum().get();
-        let valid_signers_count = self.get_action_valid_signer_count(action_id);
-        valid_signers_count >= quorum
-    }
-
-    /// The index of the last proposed action.
-    /// 0 means that no action was ever proposed yet.
-    #[view(getActionLastIndex)]
-    fn get_action_last_index(&self) -> usize {
-        self.action_mapper().len()
-    }
-
-    /// Serialized action data of an action with index.
-    #[view(getActionData)]
-    fn get_action_data(&self, action_id: usize) -> Action<Self::Api> {
-        self.action_mapper().get(action_id)
+        action_id >= min_id && action_id <= max_id
     }
 
     fn get_all_users_with_role(&self, role: UserRole) -> MultiValueEncoded<ManagedAddress> {
@@ -169,5 +94,24 @@ pub trait UtilModule: crate::storage::StorageModule {
 
         self.crypto()
             .keccak256_legacy_managed::<MAX_BUFFER_SIZE>(&serialized)
+    }
+
+    // proxies
+
+    #[proxy]
+    fn esdt_safe_proxy(&self, sc_address: ManagedAddress) -> esdt_safe::Proxy<Self::Api>;
+
+    #[proxy]
+    fn multi_transfer_esdt_proxy(
+        &self,
+        sc_address: ManagedAddress,
+    ) -> multi_transfer_esdt::Proxy<Self::Api>;
+
+    fn get_esdt_safe_proxy_instance(&self) -> esdt_safe::Proxy<Self::Api> {
+        self.esdt_safe_proxy(self.esdt_safe_address().get())
+    }
+
+    fn get_multi_transfer_esdt_proxy_instance(&self) -> multi_transfer_esdt::Proxy<Self::Api> {
+        self.multi_transfer_esdt_proxy(self.multi_transfer_esdt_address().get())
     }
 }
